use shrinkwrap for resolution
