+++ core/src/main/java/eu/fasten/core/vulchains/VulnerabilityCallChainExampleScript.java
@@ -11,6 +11,8 @@ import eu.fasten.core.data.metadatadb.codegen.tables.Modules;
 import eu.fasten.core.data.metadatadb.codegen.tables.PackageVersions;
 import eu.fasten.core.data.metadatadb.codegen.tables.Packages;
 import eu.fasten.core.maven.GraphMavenResolver;
+import eu.fasten.core.maven.ResolutionStrategy;
+import eu.fasten.core.maven.ShrinkWrapResolution;
 import eu.fasten.core.maven.data.Revision;
 import eu.fasten.core.merge.CGMerger;
 import eu.fasten.core.merge.CallGraphUtils;
@@ -90,7 +92,7 @@ public class VulnerabilityCallChainExampleScript implements Runnable {
         var metadataDb =
             CallableIndexChecker.connectToPostgres(this.metadataDbUrl, this.metadataDbUser);
         var graphResolver = buildResolver(metadataDb);
-
+        var shrinkWrapResolver = new ShrinkWrapResolution();
 
         var vulRevisions = queryVulRevisions(metadataDb);
         logger.info("Retrieved {} vulnerable packages, writing them to file.", vulRevisions.size());
@@ -114,7 +116,7 @@ public class VulnerabilityCallChainExampleScript implements Runnable {
             for (final var dependent : vulDependents.getValue()) {
                 try {
                     final var deps =
-                        resolveDeps(metadataDb, graphResolver, vulDependents, dependent);
+                        resolveDeps(metadataDb, shrinkWrapResolver, vulDependents, dependent);
 //                    final var merger = new CGMerger(ids, metadataDb, rocksDb);
                     List<ExtendedRevisionJavaCallGraph> depSet = new ArrayList<>();
                     for (Revision dep : deps) {
@@ -201,11 +203,11 @@ public class VulnerabilityCallChainExampleScript implements Runnable {
     }
 
     private Set<Revision> resolveDeps(DSLContext metadataDb,
-                                      GraphMavenResolver dependencyGraphResolver,
+                                      ResolutionStrategy dependencyGraphResolver,
                                       Map.Entry<Revision, ObjectLinkedOpenHashSet<Revision>> vulDependents,
                                       Revision dependent) {
         final var dependencySet =
-            dependencyGraphResolver.resolveDependencies(dependent, metadataDb, true);
+            dependencyGraphResolver.resolve(dependent, true);
         dependencySet.add(dependent);
         dependencySet.add(vulDependents.getKey());
         addIDsToRevisionsWithoutID(metadataDb, dependencySet);
@@ -262,7 +264,7 @@ public class VulnerabilityCallChainExampleScript implements Runnable {
     }
 
     private void addIDsToRevisionsWithoutID(
-        DSLContext metadataDb, ObjectLinkedOpenHashSet<Revision> dependencySet) {
+        DSLContext metadataDb, Set<Revision> dependencySet) {
         for (Revision revision : dependencySet) {
             if (revision.id == 0) {
                 final var id = metadataDb
